Remove deprecated pruning test and move coactivations tests to new file.
